daemon: move isTrue back to api.go as it might be generally useful
even if it isn't right now
